Abstract out how columns are rendered
